rename schema files
